add density to convert
According to the docs, you need to combine page/density. However this is
redundant information. Real page size depends on both -page AND
-density, which is strange.

Furthermore, it is strange where this 72dpi value is coming from

Included images do not depend on this 72dpi setting. Images ARE
recompressed in convert (or during extraction in pdfimags) but the
images keep the 300dpi from scanning